backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mcp
easyocr
transformers
torch
accelerate
openai
python-dotenv
requests